Internal registering of api set functional
